Extract src/configureStore.js from src/index.js
