+++ src/index.js
@@ -1,24 +1,12 @@
 // import expect from 'expect'
-import Redux, { createStore } from 'redux';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import throttle from 'lodash/throttle';
-
-import todoApp from './reducers/index';
-import { loadState, saveState } from './localStorage'
 
 import App from './components/App'
+import configureStore from './configureStore'
 
-const persistedState = loadState();
-const store = createStore(todoApp, persistedState)
-console.log(store.getState());
-
-store.subscribe(throttle(() => {
-  saveState({
-    todos: store.getState().todos
-  });
-}, 1000));
+const store = configureStore();
 
 ReactDOM.render(
   <Provider store={store}>
